requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Pillow==10.2.0
python-dotenv==1.0.1
Werkzeug==3.0.1
email-validator==2.1.0.post1
Flask-WTF==1.2.1
Flask-Migrate==4.0.5 
flask-talisman
flask-limiter
gunicorn==20.1.0
psycopg2-binary==2.9.6
cloudinary==1.32.0